port tuov_ip to pqm4

+++ crypto_sign/tuov_ip/ref/api.h
@@ -14,8 +14,7 @@
 extern  "C" {
 #endif
 
-
-
+#include <stddef.h>
 
 /** Generates a pair of public key and secret key.
  *
@@ -37,8 +36,8 @@ crypto_sign_keypair(unsigned char *pk, unsigned char *sk);
  *  \return 0 for success. -1 otherwise.
  */
 int
-crypto_sign(unsigned char *sm, unsigned long long *smlen,
-            const unsigned char *m, unsigned long long mlen,
+crypto_sign(unsigned char *sm, size_t *smlen,
+            const unsigned char *m, size_t mlen,
             const unsigned char *sk);
 
 
@@ -52,8 +51,8 @@ crypto_sign(unsigned char *sm, unsigned long long *smlen,
  *  \return 0 for successful verified. -1 for failed verification.
  */
 int
-crypto_sign_open(unsigned char *m, unsigned long long *mlen,
-                 const unsigned char *sm, unsigned long long smlen,
+crypto_sign_open(unsigned char *m, size_t *mlen,
+                 const unsigned char *sm, size_t smlen,
                  const unsigned char *pk);
 
 #ifdef __cplusplus
